[REFACTOR] new command package: mockable Runner
The idea of this change is that we don't just import `os/exec`
everywhere: for an application like this, it makes unit test coverage
difficult.

Instead, we provide a new Runner interface which abstracts the interface
for running a subprocess. This makes components behaviours much more easily
mockable, and it has the advantage that we can clearly see whether a
component has side effects based on whether it composes a Runner. This
should be done for file operations too.